Move start-account-or-login into its own dir.
