+++ frontend/lib/norent/start-account-or-login/routes.ts
@@ -1,0 +1,13 @@
+export type StartAccountOrLoginRouteInfo = ReturnType<
+  typeof createStartAccountOrLoginRouteInfo
+>;
+
+export function createStartAccountOrLoginRouteInfo(prefix: string) {
+  return {
+    phoneNumber: `${prefix}/phone/ask`,
+    verifyPhoneNumber: `${prefix}/phone/verify`,
+    setPassword: `${prefix}/password/set`,
+    verifyPassword: `${prefix}/password/verify`,
+    forgotPasswordModal: `${prefix}/password/verify/forgot-modal`,
+  };
+}
